+++ Plutarch/TermCont.hs
@@ -9,7 +9,6 @@ module Plutarch.TermCont (
   TC.pfindPlaceholder,
   pletC,
   pmatchC,
-  pletFieldsC,
   ptraceC,
   pguardC,
   pguardC',
@@ -20,12 +19,6 @@ module Plutarch.TermCont (
 import Data.Kind (Type)
 import Plutarch.Builtin.Bool (PBool, pif)
 import Plutarch.Builtin.String (PString)
-import Plutarch.DataRepr (HRec, PDataFields, PFields, pletFields)
-import Plutarch.DataRepr.Internal.Field (
-  BindFields,
-  Bindings,
-  BoundTerms,
- )
 import Plutarch.Internal.PlutusType (PlutusType, pmatch)
 import Plutarch.Internal.Term (S, Term, plet)
 import Plutarch.Internal.TermCont (TermCont, tcont)
@@ -42,18 +35,6 @@ pletC = tcont . plet
 -- | Like `pmatch` but works in a `TermCont` monad
 pmatchC :: PlutusType a => Term s a -> TermCont s (a s)
 pmatchC = tcont . pmatch
-
--- | Like `pletFields` but works in a `TermCont` monad.
-pletFieldsC ::
-  forall fs a s b ps bs.
-  ( PDataFields a
-  , ps ~ PFields a
-  , bs ~ Bindings ps fs
-  , BindFields ps bs
-  ) =>
-  Term s a ->
-  TermCont @b s (HRec (BoundTerms ps bs s))
-pletFieldsC x = tcont $ pletFields @fs x
 
 {- | Like `ptrace` but works in a `TermCont` monad.